Refactor API messages (put them into own module)
