+++ tests/api/test_root.py
@@ -2,7 +2,7 @@ from fastapi import status
 from fastapi.testclient import TestClient
 
 from run_api import app
-from src.database import schema
+from src import messages
 
 client = TestClient(app)
 
@@ -11,4 +11,4 @@ def test_get_root():
     response = client.get("/")
     assert response.status_code == status.HTTP_418_IM_A_TEAPOT
     data = response.json()
-    assert data == schema.RootMessage().model_dump()
+    assert data == messages.RootMessage().model_dump()
