+++ tests/api/test_scrape_data.py
@@ -4,7 +4,8 @@ from fastapi.testclient import TestClient
 from sqlalchemy.exc import SQLAlchemyError
 
 from run_api import app
-from src.database import get_db, schema
+from src import messages
+from src.database import get_db
 from tests.dummy_data import (
     get_test_db,
     scrape_target1,
@@ -38,7 +39,7 @@ def test_get_scrape_data_db_error(mocker):
     response = client.get("/scrape-data/")
     assert response.status_code == status.HTTP_500_INTERNAL_SERVER_ERROR
     data = response.json()
-    assert data == schema.DatabaseErrorMessage().model_dump()
+    assert data == messages.DatabaseErrorMessage().model_dump()
 
 
 def test_get_scrape_data_for_target():
@@ -53,7 +54,7 @@ def test_get_scrape_data_for_target_not_found():
     response = client.get("/scrape-data/target/99999")
     assert response.status_code == status.HTTP_404_NOT_FOUND
     data = response.json()
-    assert data == schema.TargetDoesNotExistMessage().model_dump()
+    assert data == messages.TargetDoesNotExistMessage().model_dump()
 
 
 def test_get_scrape_data_for_target_db_error(mocker):
@@ -64,7 +65,7 @@ def test_get_scrape_data_for_target_db_error(mocker):
     response = client.get(f"/scrape-data/target/{scrape_target1['id']}")
     assert response.status_code == status.HTTP_500_INTERNAL_SERVER_ERROR
     data = response.json()
-    assert data == schema.DatabaseErrorMessage().model_dump()
+    assert data == messages.DatabaseErrorMessage().model_dump()
 
 
 def test_get_scrape_data_by_id():
@@ -79,7 +80,7 @@ def test_get_scrape_data_by_id_not_found():
     response = client.get("/scrape-data/99999")
     assert response.status_code == status.HTTP_404_NOT_FOUND
     data = response.json()
-    assert data == schema.ScrapeDataDoesNotExistMessage().model_dump()
+    assert data == messages.ScrapeDataDoesNotExistMessage().model_dump()
 
 
 def test_get_scrape_data_by_id_db_error(mocker):
@@ -90,7 +91,7 @@ def test_get_scrape_data_by_id_db_error(mocker):
     response = client.get(f"/scrape-data/{scraped_data1['scrape_target_id']}")
     assert response.status_code == status.HTTP_500_INTERNAL_SERVER_ERROR
     data = response.json()
-    assert data == schema.DatabaseErrorMessage().model_dump()
+    assert data == messages.DatabaseErrorMessage().model_dump()
 
 
 def test_delete_scrape_data():
@@ -102,7 +103,7 @@ def test_delete_scrape_data_not_found():
     response = client.delete("/scrape-data/99999")
     assert response.status_code == status.HTTP_404_NOT_FOUND
     data = response.json()
-    assert data == schema.ScrapeDataDoesNotExistMessage().model_dump()
+    assert data == messages.ScrapeDataDoesNotExistMessage().model_dump()
 
 
 def test_delete_scrape_data_db_error(mocker):
@@ -113,4 +114,4 @@ def test_delete_scrape_data_db_error(mocker):
     response = client.delete(f"/scrape-data/{scraped_data1['scrape_target_id']}")
     assert response.status_code == status.HTTP_500_INTERNAL_SERVER_ERROR
     data = response.json()
-    assert data == schema.DatabaseErrorMessage().model_dump()
+    assert data == messages.DatabaseErrorMessage().model_dump()
